项目权限初始化
