+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/config/ProjectPermissionHandler.java
@@ -1,0 +1,6 @@
+package org.cloud.sonic.controller.config;
+
+public interface ProjectPermissionHandler {
+
+    int getProjectId(String method, Object[] args);
+}
