+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/controller/project/GlobalParamsController.java
@@ -15,7 +15,7 @@
  *   You should have received a copy of the GNU Affero General Public License
  *   along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.cloud.sonic.controller.controller;
+package org.cloud.sonic.controller.controller.project;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
